+++ src/pages/Main/Main.module.css
@@ -11,6 +11,14 @@
     width: 100%;
     height: 50%;
 }
+.middle_up{
+    width: 100%;
+    height: 50%;
+}
+.middle_down{
+    width: 100%;
+    height: 50%;
+}
 .footerbox{
     width: 100%;
     height: 20%;
